--- src/main/java/com/greenatom/domain/dto/ResponseDTO.java
@@ -1,19 +1,0 @@
-package com.greenatom.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * A DTO for the Response.
- */
-
-@SuppressWarnings("common-java:DuplicatedBlocks")
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ResponseDTO {
-    private String message;
-}
