+++ src/main/java/com/greenatom/domain/dto/security/JwtResponse.java
@@ -1,4 +1,4 @@
-package com.greenatom.domain.dto;
+package com.greenatom.domain.dto.security;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
